feature: move to vite. by vue cli
