+++ .eslintrc
@@ -1,5 +1,9 @@
 {
   "extends": ["plugin:vue/vue3-recommended"],
+  "env": {
+    "node": true,
+    "es2022": true
+  },
   "rules": {
     "no-unused-vars": 1,
     "vue/singleline-html-element-content-newline": "off",
