+++ package.json
@@ -2,12 +2,13 @@
   "name": "trofin.me",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
-    "serve": "vue-cli-service serve --port 3002",
+    "dev": "vite",
     "serve_api": "vercel dev --listen 3002",
-    "build": "vue-tsc --noEmit && vue-cli-service build",
-    "preview-locally": "serve -s dist",
-    "lint": "vue-cli-service lint",
+    "build": "vite build",
+    "serve": "vite preview",
+    "lint": "eslint --ext .ts,.vue,.js --ignore-path .gitignore --fix src",
     "vue_tsc:check": "vue-tsc --noEmit"
   },
   "dependencies": {
@@ -18,7 +19,6 @@
     "@fortawesome/vue-fontawesome": "^3.0.8",
     "@vercel/analytics": "^1.3.1",
     "animejs": "^3.2.2",
-    "core-js": "^3.37.1",
     "googleapis": "^137.1.0",
     "got": "^14.4.1",
     "lodash.get": "^4.4.2",
@@ -36,18 +36,15 @@
     "@types/animejs": "^3.1.12",
     "@types/node": "^20.14.10",
     "@vitejs/plugin-vue": "^5.0.5",
-    "@vue/cli-plugin-babel": "^5.0.8",
-    "@vue/cli-plugin-eslint": "^5.0.8",
-    "@vue/cli-service": "^5.0.8",
     "@vue/compiler-sfc": "^3.4.31",
     "@vue/tsconfig": "^0.5.1",
-    "babel-eslint": "^10.1.0",
     "eslint": "8.57.0",
     "eslint-plugin-vue": "^9.27.0",
     "puppeteer": "^22.13.0",
-    "tsconfig-paths-webpack-plugin": "^4.1.0",
     "typescript": "^5.5.3",
+    "vite": "^5.3.3",
     "vite-plugin-checker": "^0.7.1",
+    "vite-tsconfig-paths": "^4.3.2",
     "vue-svg-loader": "^0.16.0",
     "vue-tsc": "^2.0.26"
   },
@@ -60,9 +57,6 @@
       "plugin:vue/vue3-essential",
       "eslint:recommended"
     ],
-    "parserOptions": {
-      "parser": "babel-eslint"
-    },
     "rules": {}
   },
   "browserslist": [
